basic parcel setup using es6 modules
